Move FallBackMaxReorg and create tests
